Refactoring - pulling static DNS stuff into its own file
